Split into conanfile_base.py and conanfile.py
