+++ conanfile_base.py
@@ -1,0 +1,52 @@
+from conans import ConanFile, CMake, tools
+import os
+
+
+class ConanFileBase(ConanFile):
+    _base_name = "protobuf"
+    version = "3.6.1"
+    description = "Protocol Buffers - Google's data interchange format"
+    topics = ("conan", "protobuf", "protocol-buffers", "protocol-compiler", "serialization", "rpc")
+    url = "https://github.com/bincrafters/conan-protobuf"
+    homepage = "https://github.com/protocolbuffers/protobuf"
+    author = "Bincrafters <bincrafters@gmail.com>"
+    license = "BSD-3-Clause"
+    exports = ["LICENSE.md"]
+    exports_sources = ["CMakeLists.txt", "protobuf.patch"]
+    generators = "cmake"
+    short_paths = True
+
+    # Custom attributes for Bincrafters recipe conventions
+    _source_subfolder = "source_subfolder"
+    _build_subfolder = "build_subfolder"
+
+    @property
+    def _is_clang_x86(self):
+        return self.settings.compiler == "clang" and self.settings.arch == "x86"
+
+    def source(self):
+        sha256 = "3d4e589d81b2006ca603c1ab712c9715a76227293032d05b26fca603f90b3f5b"
+        tools.get("{0}/archive/v{1}.tar.gz".format(self.homepage, self.version), sha256=sha256)
+        extracted_dir = self.name + "-" + self.version
+        os.rename(extracted_dir, self._source_subfolder)
+
+    def _configure_cmake(self):
+        cmake = CMake(self, set_cmake_flags=True)
+        cmake.definitions["protobuf_BUILD_TESTS"] = False
+        cmake.definitions["protobuf_WITH_ZLIB"] = self.options.with_zlib
+        cmake.definitions["protobuf_BUILD_PROTOC_BINARIES"] = not self.options.lite
+        cmake.definitions["protobuf_BUILD_PROTOBUF_LITE"] = self.options.lite
+        if self.settings.compiler == "Visual Studio":
+            cmake.definitions["protobuf_MSVC_STATIC_RUNTIME"] = "MT" in self.settings.compiler.runtime
+        cmake.configure(build_folder=self._build_subfolder)
+        return cmake
+
+    def build(self):
+        tools.patch(base_path=self._source_subfolder, patch_file="protobuf.patch")
+        cmake = self._configure_cmake()
+        cmake.build()
+
+    def package(self):
+        self.copy("LICENSE", dst="licenses", src=self._source_subfolder)
+        cmake = self._configure_cmake()
+        cmake.install()
